# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BaleBeleqNews API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  color: string;
  articleCount?: number;
}

export interface Article {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  imageCaption?: string | null;
  /** @nullable */
  categoryId?: number | null;
  category?: Category | null;
  isPublished: boolean;
  isFeatured: boolean;
  viewCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ArticleListResponse {
  articles: Article[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type ArticleStatsByCategoryItem = {
  categoryName: string;
  count: number;
};

export interface ArticleStats {
  total: number;
  published: number;
  draft: number;
  totalViews: number;
  byCategory: ArticleStatsByCategoryItem[];
}

export interface ArticleInput {
  title: string;
  excerpt: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  imageCaption?: string | null;
  /** @nullable */
  categoryId?: number | null;
  isPublished: boolean;
  isFeatured: boolean;
}

export interface CategoryInput {
  name: string;
  slug: string;
  color: string;
}

export interface LoginCredentials {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
  /** @nullable */
  email?: string | null;
  role?: string;
}

export interface LoginResponse {
  token: string;
  user: AdminUser;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export type RequestUploadUrlResponseMetadata = {
  name?: string;
  size?: number;
  contentType?: string;
};

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: RequestUploadUrlResponseMetadata;
}

export interface AdminUserCreate {
  username: string;
  password: string;
  email?: string;
  role: string;
}

export type ListArticlesParams = {
page?: number;
limit?: number;
categoryId?: number;
search?: string;
};

export type ListAllArticlesParams = {
page?: number;
limit?: number;
};