# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BaleBeleqNews API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: articles
    description: News article operations
  - name: categories
    description: Category operations
  - name: auth
    description: Admin authentication
  - name: storage
    description: File upload operations
  - name: users
    description: Admin user management (owner only)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List published articles
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create a new article (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Article created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized

  /admin/articles/{id}:
    get:
      operationId: getAdminArticle
      tags: [articles]
      summary: Get any article by ID including drafts (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /admin/articles:
    get:
      operationId: listAllArticles
      tags: [articles]
      summary: List all articles including drafts (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: All articles (published + draft)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"
        "401":
          description: Unauthorized

  /articles/featured:
    get:
      operationId: getFeaturedArticles
      tags: [articles]
      summary: Get featured/latest articles for homepage
      responses:
        "200":
          description: Featured articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"

  /articles/stats:
    get:
      operationId: getArticleStats
      tags: [articles]
      summary: Get article statistics (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Article statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleStats"

  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get single article by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
    put:
      operationId: updateArticle
      tags: [articles]
      summary: Update article (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "200":
          description: Article updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete article (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

  /articles/{id}/slug:
    get:
      operationId: getArticleBySlug
      tags: [articles]
      summary: Get single article by slug
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found

  /articles/{id}/views:
    post:
      operationId: incrementArticleViews
      tags: [articles]
      summary: Increment article view count
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: View incremented

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request presigned upload URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
        "401":
          description: Unauthorized

  /storage/objects/{objectPath}:
    get:
      operationId: getObject
      tags: [storage]
      summary: Get/redirect to object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content or redirect
        "404":
          description: Not found

  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [users]
      summary: List all admin users (owner only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of admin users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (owner only)
    post:
      operationId: createAdminUser
      tags: [users]
      summary: Create admin user (owner only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserCreate"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (owner only)

  /admin/users/{id}:
    put:
      operationId: updateAdminUser
      tags: [users]
      summary: Update admin user (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserCreate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (owner only)
    delete:
      operationId: deleteAdminUser
      tags: [users]
      summary: Delete admin user (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (owner only)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Article:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        imageCaption:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        category:
          oneOf:
            - $ref: "#/components/schemas/Category"
            - type: "null"
        isPublished:
          type: boolean
        isFeatured:
          type: boolean
        viewCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - slug
        - excerpt
        - content
        - isPublished
        - isFeatured
        - viewCount
        - createdAt
        - updatedAt

    ArticleListResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - articles
        - total
        - page
        - limit
        - totalPages

    ArticleStats:
      type: object
      properties:
        total:
          type: integer
        published:
          type: integer
        draft:
          type: integer
        totalViews:
          type: integer
        byCategory:
          type: array
          items:
            type: object
            properties:
              categoryName:
                type: string
              count:
                type: integer
            required:
              - categoryName
              - count
      required:
        - total
        - published
        - draft
        - totalViews
        - byCategory

    ArticleInput:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        imageCaption:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        isPublished:
          type: boolean
        isFeatured:
          type: boolean
      required:
        - title
        - excerpt
        - content
        - isPublished
        - isFeatured

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        color:
          type: string
        articleCount:
          type: integer
      required:
        - id
        - name
        - slug
        - color

    CategoryInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        color:
          type: string
      required:
        - name
        - slug
        - color

    LoginCredentials:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
      required:
        - token
        - user

    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
      required:
        - id
        - username

    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType

    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            size:
              type: integer
            contentType:
              type: string
      required:
        - uploadURL
        - objectPath

    AdminUserCreate:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
        role:
          type: string
      required:
        - username
        - password
        - role